chore: refactor non test classes
