+++ buildSrc/src/main/kotlin/io.customer/android/Dependencies.kt
@@ -32,8 +32,8 @@ object Dependencies {
         "io.github.gradle-nexus:publish-plugin:${Versions.GRADLE_NEXUS_PUBLISH_PLUGIN}"
     const val junit4 = "junit:junit:${Versions.JUNIT4}"
     const val junitBom = "org.junit:junit-bom:${Versions.JUNIT_BOM}"
-    const val junitEngine = "org.junit.vintage:junit-vintage-engine"
     const val junitJupiter = "org.junit.jupiter:junit-jupiter"
+    const val junitVintageEngine = "org.junit.vintage:junit-vintage-engine"
     const val kluent = "org.amshove.kluent:kluent-android:${Versions.KLUENT}"
     const val kotlinBinaryValidator =
         "org.jetbrains.kotlinx:binary-compatibility-validator:${Versions.KOTLIN_BINARY_VALIDATOR}"
