+++ datapipelines/build.gradle
@@ -50,8 +50,8 @@ dependencies {
     testImplementation(platform(Dependencies.junitBom))
     testImplementation(Dependencies.junitJupiter)
 
-    // enables running JUnit 4 tests within the JUnit 5 (jupitor) platform.
-    testRuntimeOnly(Dependencies.junitEngine)
+    // enables running JUnit 4 tests within the JUnit 5 (jupiter) platform.
+    testRuntimeOnly(Dependencies.junitVintageEngine)
 }
 
 tasks.withType(Test).configureEach {
