+++ datapipelines/src/test/java/io/customer/datapipelines/DataPipelinesCompatibilityTests.kt
@@ -4,14 +4,14 @@ import com.segment.analytics.kotlin.core.Storage
 import com.segment.analytics.kotlin.core.emptyJsonArray
 import com.segment.analytics.kotlin.core.emptyJsonObject
 import com.segment.analytics.kotlin.core.utilities.getString
-import io.customer.datapipelines.core.UnitTest
-import io.customer.datapipelines.extensions.decodeJson
-import io.customer.datapipelines.extensions.deviceToken
-import io.customer.datapipelines.extensions.encodeToJsonElement
-import io.customer.datapipelines.extensions.shouldMatchTo
-import io.customer.datapipelines.extensions.toJsonObject
-import io.customer.datapipelines.support.UserTraits
-import io.customer.datapipelines.utils.TestConstants
+import io.customer.datapipelines.support.core.UnitTest
+import io.customer.datapipelines.support.data.model.UserTraits
+import io.customer.datapipelines.support.extensions.decodeJson
+import io.customer.datapipelines.support.extensions.deviceToken
+import io.customer.datapipelines.support.extensions.encodeToJsonElement
+import io.customer.datapipelines.support.extensions.shouldMatchTo
+import io.customer.datapipelines.support.extensions.toJsonObject
+import io.customer.datapipelines.support.utils.TestConstants
 import io.customer.sdk.CustomerIO
 import io.customer.sdk.CustomerIOBuilder
 import io.customer.sdk.data.model.CustomAttributes
