+++ datapipelines/src/test/java/io/customer/datapipelines/support/extensions/JsonExt.kt
@@ -1,4 +1,4 @@
-package io.customer.datapipelines.extensions
+package io.customer.datapipelines.support.extensions
 
 import com.segment.analytics.kotlin.core.emptyJsonObject
 import io.customer.sdk.data.model.CustomAttributes
@@ -12,13 +12,13 @@ import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.encodeToJsonElement
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
-import org.amshove.kluent.internal.assertEquals
+import org.amshove.kluent.shouldBeEqualTo
 
 /**
  * Similar to Kluent's `shouldBeEqualTo` but for comparing JSON objects with custom attributes map.
  */
 infix fun JsonObject.shouldMatchTo(expected: CustomAttributes): JsonObject {
-    return this.apply { assertEquals(expected.toJsonObject(), this) }
+    return this.apply { this shouldBeEqualTo expected.toJsonObject() }
 }
 
 /**
